requirements.txt
cx_Freeze==6.15.10
Pillow==10.0.1
watchdog==3.0.0